pyproject.toml
[tool.poetry]
name = "postgresql-geometry"
version = "0.1.0"
description = "PostgreSQL geometry types for python"
authors = ["Thomas Berdy <thomas.berdy@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.django]
optional = true

[tool.poetry.group.django.dependencies]
Django = ">=4.2"

[tool.poetry.group.factoryboy]
optional = true

[tool.poetry.group.factoryboy.dependencies]
factory-boy = ">=3.3.0"

[tool.poetry.group.faker]
optional = true

[tool.poetry.group.faker.dependencies]
Faker = ">=25.0.0"

[tool.poetry.group.rest_framework]
optional = true

[tool.poetry.group.rest_framework.dependencies]
djangorestframework = ">=3.14.0"

[tool.poetry.group.sqlalchemy]
optional = true

[tool.poetry.group.sqlalchemy.dependencies]
SQLAlchemy = ">=2.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
psycopg = "~3.1.18"
pytest = "~8.2.0"
pytest-cov = "~5.0.0"
pytest-django = "~4.8.0"
pytest-env = "~1.1.3"
pytest-socket = "~0.7.0"
pytest-sugar = "~1.0.0"
python-dotenv = "~1.0.1"
ruff = "~0.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "N",
    "YTT",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EXE",
    "FA",
    "ISC",
    "ICN",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "PTH",
    "TD",
    "PERF",
    # "FURB",
    "LOG",
    "RUF",
]
ignore = [
    "COM812",
    "ISC001",
    # TODO: enable this
    "TCH001",
    "TCH002",
    "TCH003",
    "TD002",
    "TD003",
    "RUF012",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "postgresql_geometry_tests.django.settings"
addopts = [
    "--cov=.",
    "--disable-socket",
    "--allow-unix-socket",
]
python_files = [
    "test_*.py",
]
testpaths = [
    "postgresql_geometry_tests",
]

[tool.coverage.report]
include = [
    "postgresql_geometry/**/*.py",
]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING",
    "@abstractmethod",
]
